+++ src/aws_encryption_sdk_cli/internal/identifiers.py
@@ -30,7 +30,7 @@ __all__ = (
     'DEFAULT_MASTER_KEY_PROVIDER',
     'OperationResult'
 )
-__version__ = '1.1.0'  # type: str
+__version__ = '1.1.1'  # type: str
 
 #: Suffix added to output files if specific output filename is not specified.
 OUTPUT_SUFFIX = {